requirements.txt
jira
urllib3<2
python-dotenv
python-dateutil
pandas
matplotlib
pytest
atlassian-python-api
markdown2